feat(rpc): 增强RPC框架示例及文档，完善服务调用流程
- 为User实体类添加详细JavaDoc，说明序列化关键点和Lombok用法
- 补充UserService接口注释，解释RPC调用机制及接口在common模块的作用
- 实现UserService的静态代理UserServiceProxy，展示RPC调用的序列化及网络传输流程
- 完善ConsumerExample示例，展示基于动态代理的RPC调用流程及负载均衡验证
- 新增EasyConsumerExample示例，演示静态代理的手写实现，方便理解代理原理
- 增加EasyProviderExample示例，手动注册服务及启动HTTP服务器，助力框架原理学习
- 更新ProviderExample，展示服务注册列表和自动启动流程，接近生产环境用法
- UserServiceImpl实现中打印服务端口，方便调试负载均衡
- 在Spring Boot消费者示例中用@RpcReference注解实现远程服务注入，简化调用
- 添加ExampleSpringbootConsumerApplication启动类，说明消费者不启动服务监听端口
- 编写Spring Boot消费者端集成测试，确保RPC调用链路和容器上下文正常加载
- 补充Spring Boot提供者启动类注释，突出自动化配置优势及与原生API对比

+++ hgl-rpc-core/src/main/java/com/hgl/hglrpc/protocol/ProtocolConstant.java
@@ -1,26 +1,56 @@
 package com.hgl.hglrpc.protocol;
 
 /**
- * @ClassName: ProtocolConstant
- * @Package: com.hgl.hglrpc.protocol
- * @Description: 协议常量
+ * 协议常量 —— "交通规则的基石"
+ *
+ * <p>自定义二进制协议就像一套交通规则，规定了数据在网络中"行驶"的格式。
+ * ProtocolConstant 定义了这套规则中最基本的三个常量，
+ * 所有编解码器（Encoder/Decoder）都必须严格遵守。
+ *
+ * <p>协议报文结构（共 17 字节头 + 变长体）：
+ * <pre>
+ *   ┌─────────┬─────────┬────────────┬────────┬────────┬────────────┬────────────┬───────────┐
+ *   │ magic   │ version │ serializer │  type  │ status │ requestId  │ bodyLength │   body    │
+ *   │ 1 byte  │ 1 byte  │  1 byte    │ 1 byte │ 1 byte │  8 bytes   │  4 bytes   │ N bytes   │
+ *   └─────────┴─────────┴────────────┴────────┴────────┴────────────┴────────────┴───────────┘
+ *   |&lt;----------- 17 字节消息头 ----------&gt;|&lt;------------ 变长消息体 ------------&gt;|
+ * </pre>
+ *
+ * <p>为什么需要魔数？
+ * 就像人民币上的水印，魔数用于快速鉴别"这包数据是不是我认识的协议"。
+ * 如果收到的数据第一个字节不是魔数，说明要么是垃圾数据，要么是别的协议，
+ * 可以直接丢弃，避免浪费 CPU 去解析无意义的内容。
+ *
  * @Author HGL
  * @Create: 2025/9/3 14:56
  */
 public interface ProtocolConstant {
 
     /**
-     * 消息头长度
+     * 消息头长度 = 17 字节
+     *
+     * <p>组成：magic(1) + version(1) + serializer(1) + type(1) + status(1) + requestId(8) + bodyLength(4)
+     * 解码器在读取数据时，先读 17 字节解析出头部，再根据 bodyLength 读取剩余的消息体。
+     * 这就是解决 TCP 粘包/拆包问题的核心思路——"定长头 + 变长体"。
      */
     int MESSAGE_HEADER_LENGTH = 17;
 
     /**
-     * 协议魔数
+     * 协议魔数 —— 协议的"指纹"
+     *
+     * <p>值为 0x1（十六进制的 1）。
+     * 实际生产环境中，建议使用更独特的值（如 0xCA），避免与其他协议巧合冲突。
+     * 类比：Dubbo 的魔数是 0xdabb，HTTP 的魔数是 "GET"/"POST" 等文本前缀。
      */
-    byte PROTOCOL_MAGIC = 0x1;
+    int PROTOCOL_MAGIC = 0x1;
 
     /**
-     * 协议版本号
+     * 协议版本号 —— 协议的"版本标记"
+     *
+     * <p>值为 0x1，表示当前是第一版协议。
+     * 未来如果协议格式有变（比如增加新的头部字段），
+     * 可以通过版本号来区分新旧协议，实现向后兼容。
+     * 类比：HTTP/1.0 vs HTTP/1.1 vs HTTP/2.0。
      */
-    byte PROTOCOL_VERSION = 0x1;
+    int PROTOCOL_VERSION = 0x1;
 }
